fix: make deepName public and move to extensions package
It's a very handy utility method.
