+++ src/main/kotlin/com/expedia/graphql/schema/extensions/graphQLTypeExtensions.kt
@@ -1,21 +1,9 @@
 package com.expedia.graphql.schema.extensions
 
 import com.expedia.graphql.exceptions.NestingNonNullTypeException
-import graphql.schema.GraphQLList
 import graphql.schema.GraphQLNonNull
 import graphql.schema.GraphQLType
 import kotlin.reflect.KType
-
-/**
- * Renders a readable string from the given graphql type no matter how deeply nested
- * Eg: [[Int!]]!
- */
-internal val GraphQLType.deepName: String
-    get() = when {
-        this is GraphQLNonNull -> "${this.wrappedType.deepName}!"
-        this is GraphQLList -> "[${this.wrappedType.deepName}]"
-        else -> name
-    }
 
 /**
  * Map null and non-null types.
