+++ src/test/kotlin/com/expedia/graphql/schema/dataFetchers/CustomDataFetcherTests.kt
@@ -2,7 +2,7 @@ package com.expedia.graphql.schema.dataFetchers
 
 import com.expedia.graphql.TopLevelObjectDef
 import com.expedia.graphql.schema.SchemaGeneratorConfig
-import com.expedia.graphql.schema.extensions.deepName
+import com.expedia.graphql.extensions.deepName
 import com.expedia.graphql.toSchema
 import graphql.GraphQL
 import graphql.schema.DataFetcher
