+++ src/test/kotlin/com/expedia/graphql/schema/extensions/GraphQLTypeExtensionsTest.kt
@@ -1,6 +1,7 @@
 package com.expedia.graphql.schema.extensions
 
 import com.expedia.graphql.exceptions.NestingNonNullTypeException
+import com.expedia.graphql.extensions.deepName
 import graphql.schema.GraphQLList
 import graphql.schema.GraphQLNonNull
 import graphql.schema.GraphQLType
